PERA-2101 Separate account and asset list items
